pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bigbob-verify-bot"
version = "0.1.0"
description = "Telegram bot and webhook service for Bigbob verification"
authors = [{name = "Bigbob"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot[ext]>=20.4",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.27",
    "psycopg[binary]>=3.1",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "redis>=5.0",
    "aiohttp>=3.9",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
    "alembic>=1.12",
    "httpx>=0.27",
    "pytest>=7.4",
    "pytest-asyncio>=0.21"
]

[tool.pytest.ini_options]
testpaths = ["bot/tests"]